+++ circuit-foundation/src/commonJvmTest/kotlin/com/slack/circuit/foundation/NavigatorTest.kt
@@ -19,7 +19,7 @@ class NavigatorTestUtil {
   @Test
   fun popAtRoot() {
     val backStack = SaveableBackStack(TestScreen)
-    backStack.push(TestScreen)
+    backStack.push(TestScreen2)
 
     var onRootPop = 0
     val navigator = NavigatorImpl(backStack) { onRootPop++ }